feat: inject js split providers
