+++ src/pageProvider/index.js
@@ -1,375 +1,31 @@
 import { ProviderManager } from './providerManager'
+import { addEthereumChain } from './ethereumProvider'
+import { addGlobalEthereumProvider } from './globalEthereumProvider'
+import { addBitcoinProvider } from './bitcoinProvider'
+import { addPaymentUriProvider } from './paymentUri'
+import { addTerraProvider } from './terraProvider'
+import { addNearProvider } from './nearProvider'
+import { addSolanaProvider } from './solanaProvider'
+
+// TODO: refactor providers here to follow a more opiniated format
+// class PageProvider {
+//   setup() {}
+// }
+// class BitcoinProvider extends PageProvider {
+//   setup(window) {
+//     window.bitcoin = {
+//       isLiquality: true
+//     }
+//   }
+// }
+// const providers = [new BitcoinProvider()]
+// providers.forEach((p) => p.setup())
 
 window.providerManager = new ProviderManager()
-
-const metamaskEmulated = ['opensea.io', 'unstoppabledomains.com'].some(
-  (site) => window.location.host.indexOf(site) !== -1
-) // Is some kind of smart emulation possible?
-
-function addEthereumChain({ chain, asset, network }) {
-  const injectionName = window.providerManager.getInjectionName(chain)
-
-  async function getAddresses() {
-    const eth = window.providerManager.getProviderFor(asset)
-    let addresses = await eth.getMethod('wallet.getAddresses')()
-    addresses = addresses.map((a) => a.address)
-    window[injectionName].selectedAddress = addresses[0]
-    return addresses
-  }
-
-  async function ethereumHandleRequest(req) {
-    const eth = window.providerManager.getProviderFor(asset)
-    if (req.method.startsWith('metamask_')) return null
-
-    if (req.method === 'eth_requestAccounts') {
-      return await window[injectionName].enable(chain)
-    }
-    if (req.method === 'personal_sign') {
-      const sig = await eth.getMethod('wallet.signMessage')(req.params[0], req.params[1])
-      return '0x' + sig
-    }
-
-    if (
-      req.method === 'eth_signTypedData' ||
-      req.method === 'eth_signTypedData_v3' ||
-      req.method === 'eth_signTypedData_v4'
-    ) {
-      const sig = await eth.getMethod('wallet.signTypedData')(req)
-      return sig
-    }
-
-    if (req.method === 'eth_sendTransaction') {
-      const to = req.params[0].to
-      const value = req.params[0].value
-      const data = req.params[0].data
-      const gas = req.params[0].gas
-      const result = await eth.getMethod('wallet.sendTransaction')({ to, value, data, gas })
-      return result.hash
-    }
-    if (req.method === 'eth_accounts') {
-      return getAddresses()
-    }
-    return eth.getMethod('chain.sendRpcRequest')(req.method, req.params)
-  }
-
-  window[injectionName] = {
-    isLiquality: true,
-    isMetaMask: metamaskEmulated,
-    isEIP1193: true,
-    networkVersion: network.networkId,
-    chainId: network.chainId.toString(16),
-    enable: async () => {
-      const { accepted, _chain } = await window.providerManager.enable(chain)
-      if (!accepted) throw new Error('User rejected')
-      return getAddresses()
-    },
-    request: async (req) => {
-      const params = req.params || []
-      return ethereumHandleRequest({
-        method: req.method,
-        params
-      })
-    },
-    send: async (req, _paramsOrCallback) => {
-      if (typeof _paramsOrCallback === 'function') {
-        window[injectionName].sendAsync(req, _paramsOrCallback)
-        return
-      }
-      const method = typeof req === 'string' ? req : req.method
-      const params = req.params || _paramsOrCallback || []
-      return ethereumHandleRequest({ method, params }).then((result) => ({
-        id: req && req.id ? req.id : 99999,
-        jsonrpc: '2.0',
-        result
-      }))
-    },
-    removeAllListeners: (event) => {
-      // mock this call for a hotfix
-      return false
-    },
-    sendAsync: (req, callback) => {
-      return ethereumHandleRequest(req)
-        .then((result) =>
-          callback(null, {
-            id: req.id,
-            jsonrpc: '2.0',
-            result
-          })
-        )
-        .catch((err) => callback(err))
-    },
-    on: (method, callback) => {
-      if (method === 'chainChanged') {
-        window.addEventListener('liqualityChainChanged', ({ detail }) => {
-          const result = JSON.parse(detail)
-          callback('0x' + result.chainIds[chain].toString(16))
-        })
-      }
-
-      if (method === 'accountsChanged') {
-        window.addEventListener('liqualityAccountsChanged', () => {
-          const addresses = getAddresses()
-          callback(addresses)
-        })
-      }
-    },
-    autoRefreshOnNetworkChange: false
-  }
-}
-
 window.liquality.evmChains.forEach((evmChain) => addEthereumChain(evmChain))
-// chain, override
-function proxyEthereum(chain) {
-  window.ethereumProxyChain = chain
-
-  const enable = async () => {
-    const { accepted, chain } = await window.providerManager.enable()
-    if (!accepted) throw new Error('User rejected')
-    const chainChanged = window.ethereumProxyChain !== chain
-    window.ethereumProxyChain = chain
-    const injectionName = window.providerManager.getInjectionName(chain)
-    if (chainChanged) {
-      window.dispatchEvent(
-        new CustomEvent('liqualityChainChanged', {
-          detail: JSON.stringify({ chainIds: { [chain]: window[injectionName].chainId } })
-        })
-      )
-    }
-    return window[injectionName].enable(chain)
-  }
-
-  const overrideHandler = {
-    get: function (_target, prop, receiver) {
-      const injectionName = window.providerManager.getInjectionName(window.ethereumProxyChain)
-      const target = window[injectionName]
-
-      if (prop === 'on') {
-        return (method, callback) => {
-          if (method === 'chainChanged') {
-            window.addEventListener('liqualityChainChanged', ({ detail }) => {
-              const result = JSON.parse(detail)
-              callback('0x' + result.chainIds[window.ethereumProxyChain].toString(16))
-            })
-          }
-
-          if (method === 'accountsChanged') {
-            window.addEventListener('liqualityAccountsChanged', () => {
-              target.request({ method: 'eth_accounts', params: [] }).then((newAccounts) => {
-                callback(newAccounts)
-              })
-            })
-          }
-        }
-      }
-
-      if (prop === 'enable') {
-        return async () => {
-          return enable()
-        }
-      }
-
-      if (prop === 'request') {
-        return async (req) => {
-          if (req.method === 'eth_requestAccounts') {
-            return enable()
-          }
-          return target[prop](req)
-        }
-      }
-
-      if (prop === 'send') {
-        return async (req, _paramsOrCallback) => {
-          const method = typeof req === 'string' ? req : req.method
-          if (method === 'eth_requestAccounts') {
-            if (typeof _paramsOrCallback === 'function') {
-              const callback = _paramsOrCallback
-              return enable()
-                .then((result) =>
-                  callback(null, {
-                    id: req.id,
-                    jsonrpc: '2.0',
-                    result
-                  })
-                )
-                .catch((err) => callback(err))
-            }
-            return enable()
-          }
-          return target[prop](req, _paramsOrCallback)
-        }
-      }
-
-      if (prop === 'sendAsync') {
-        // return async (req, _paramsOrCallback) => {
-        //   window.req = req
-        //   // if(req.method === 'wallet_switchEthereumChain') {
-        //   //   const { params } = req
-        //   //   const { chainId } = params[0]
-        //   //   if(chainId === "0x89") {
-        //   //     const callback = _paramsOrCallback
-        //   //     await window.providerManager.enable('polygon', true)
-        //   //     return callback(null, {
-        //   //       id: req.id,
-        //   //       jsonrpc: '2.0',
-        //   //       result: null
-        //   //     })
-        //   //   }
-        //   //}
-        //   return target[prop](req, _paramsOrCallback)
-        // }
-      }
-
-      return target[prop]
-    }
-  }
-  window.ethereum = new Proxy({}, overrideHandler)
-}
-
-function overrideEthereum(chain) {
-  proxyEthereum(chain)
-}
-
-const { override, ethereumChain } = window.liquality.globalEthereum
-
-if (override) {
-  overrideEthereum(ethereumChain)
-  const retryLimit = 5
-  let retries = 0
-  const interval = setInterval(() => {
-    retries++
-    if (window.ethereum && !window.ethereum.isLiquality) {
-      overrideEthereum(ethereumChain)
-      clearInterval(interval)
-    }
-    if (retries >= retryLimit) clearInterval(interval)
-  }, 1000)
-} else if (!window.ethereum) {
-  overrideEthereum(ethereumChain)
-}
-
-const BITCOIN_REQUEST_MAP = {
-  wallet_getConnectedNetwork: 'wallet.getConnectedNetwork',
-  wallet_getAddresses: 'wallet.getAddresses',
-  wallet_signMessage: 'wallet.signMessage',
-  wallet_sendTransaction: 'wallet.sendTransaction',
-  wallet_signPSBT: 'wallet.signPSBT'
-}
-
-async function bitcoinHandleRequest(req) {
-  const btc = window.providerManager.getProviderFor('BTC')
-  if (req.method === 'wallet_sendTransaction') {
-    const to = req.params[0].to
-    const value = req.params[0].value.toString(16)
-    return btc.getMethod('wallet.sendTransaction')({ to, value })
-  }
-  const method = BITCOIN_REQUEST_MAP[req.method] || req.method
-  return btc.getMethod(method)(...req.params)
-}
-
-window.bitcoin = {
-  enable: async () => {
-    const { accepted } = await window.providerManager.enable('bitcoin')
-    if (!accepted) throw new Error('User rejected')
-    const btc = window.providerManager.getProviderFor('BTC')
-    return btc.getMethod('wallet.getAddresses')()
-  },
-  request: async (req) => {
-    const params = req.params || []
-    return bitcoinHandleRequest({
-      method: req.method,
-      params
-    })
-  }
-}
-
-const COMMON_REQUEST_MAP = {
-  wallet_getConnectedNetwork: 'wallet.getConnectedNetwork',
-  wallet_getAddresses: 'wallet.getAddresses',
-  wallet_signMessage: 'wallet.signMessage',
-  wallet_sendTransaction: 'wallet.sendTransaction'
-}
-
-async function nearHandleRequest(req) {
-  const near = window.providerManager.getProviderFor('NEAR')
-  const method = COMMON_REQUEST_MAP[req.method] || req.method
-  return near.getMethod(method)(...req.params)
-}
-window.near = {
-  enable: async () => {
-    const { accepted } = await window.providerManager.enable('near')
-    if (!accepted) throw new Error('User rejected')
-    const near = window.providerManager.getProviderFor('NEAR')
-    return near.getMethod('wallet.getAddresses')()
-  },
-  request: async (req) => {
-    const params = req.params || []
-    return nearHandleRequest({
-      method: req.method,
-      params
-    })
-  }
-}
-
-async function solanaHandleRequest(req) {
-  const solana = window.providerManager.getProviderFor('SOL')
-  const method = COMMON_REQUEST_MAP[req.method] || req.method
-  return solana.getMethod(method)(...req.params)
-}
-window.sollet = {
-  enable: async () => {
-    const { accepted } = await window.providerManager.enable('solana')
-    if (!accepted) throw new Error('User rejected')
-    const solana = window.providerManager.getProviderFor('SOL')
-    return solana.getMethod('wallet.getAddresses')()
-  },
-  request: async (req) => {
-    const params = req.params || []
-    return solanaHandleRequest({
-      method: req.method,
-      params
-    })
-  }
-}
-
-async function terraHandleRequest(req) {
-  const terraProvider = window.providerManager.getProviderFor('LUNA')
-  const method = COMMON_REQUEST_MAP[req.method] || req.method
-  return terraProvider.getMethod(method)(...req.params)
-}
-window.isTerraExtensionAvailable = true
-window.terra = {
-  enable: async () => {
-    const accepted = await window.providerManager.enable('terra')
-    if (!accepted) throw new Error('User rejected')
-    const terra = window.providerManager.getProviderFor('LUNA')
-    return terra.getMethod('wallet.getAddresses')()
-  },
-  request: async (req) => {
-    const params = req.params || []
-    return terraHandleRequest({
-      method: req.method,
-      params
-    })
-  }
-}
-
-document.addEventListener(
-  'DOMContentLoaded',
-  () => {
-    document.body.addEventListener('click', async (e) => {
-      const element = e.target
-      if (!element || !element.closest) return
-      const uri = element.closest('[href^="bitcoin:"]') || element.closest('[href^="ethereum:"]')
-      if (uri) {
-        const href = uri.getAttribute('href')
-        const includesAmount = href.includes('value=') || href.includes('amount=')
-        if (includesAmount) {
-          e.preventDefault()
-          await window.providerManager.enable('near')
-          window.providerManager.proxy('HANDLE_PAYMENT_URI', { uri: href })
-        }
-      }
-    })
-  },
-  { once: true }
-)
+addGlobalEthereumProvider()
+addBitcoinProvider()
+addNearProvider()
+addSolanaProvider()
+addTerraProvider()
+addPaymentUriProvider()
