+++ src/pageProvider/utils.js
@@ -1,0 +1,8 @@
+const COMMON_REQUEST_MAP = {
+  wallet_getConnectedNetwork: 'wallet.getConnectedNetwork',
+  wallet_getAddresses: 'wallet.getAddresses',
+  wallet_signMessage: 'wallet.signMessage',
+  wallet_sendTransaction: 'wallet.sendTransaction'
+}
+
+export { COMMON_REQUEST_MAP }
